hack: update format of rhcos build metadata
This splits the RHCOS build metadata into architecture-specific files.
This will allow the metadata to contain information about bootimages of
multiple architectures. In order to preserve backward compatibility
(there are a few users, including certain CI jobs, that pull rhcos.json
from GitHub directly), I've opted to use seperate files and to symlink
the original file to the AMD64 variant.

When updating the RHCOS bootimages, this script will need to be run once
for each architecture that is being updated.